refactor: introduce application layer in Run
