+++ src/php/Run/RunFactory.php
@@ -10,12 +10,12 @@ use Phel\Command\CommandFacadeInterface;
 use Phel\Compiler\CompilerFacadeInterface;
 use Phel\Printer\Printer;
 use Phel\Printer\PrinterInterface;
+use Phel\Run\Application\NamespaceRunner;
 use Phel\Run\Domain\Repl\ColorStyle;
 use Phel\Run\Domain\Repl\ColorStyleInterface;
 use Phel\Run\Domain\Repl\ReplCommandIoInterface;
 use Phel\Run\Domain\Repl\ReplCommandSystemIo;
 use Phel\Run\Domain\Runner\NamespaceCollector;
-use Phel\Run\Domain\Runner\NamespaceRunner;
 use Phel\Run\Domain\Runner\NamespaceRunnerInterface;
 
 /**
